Created the Classes for CreatePortfolioActivity. Testing pending.

+++ template_java_project/src/com/cryptoportfolio/converter/ModelConverter.java
@@ -13,7 +13,7 @@ public class ModelConverter {
      */
     public PortfolioModel toPortfolioModel(String username, UserPortfolio userAssetsMapping) {
 
-        Map<String, Integer> portfolioAssetsMap = userAssetsMapping.getAssetQuantityMap();
+        Map<String, Double> portfolioAssetsMap = userAssetsMapping.getAssetQuantityMap();
 
         return PortfolioModel.builder()
                 .withUsername(username)
